requirements.txt
streamlit==1.29.0
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.0
plotly==5.18.0
matplotlib==3.8.4
seaborn==0.13.0
xgboost==2.0.3
joblib==1.4.2
